Rename pokemon actions
